qiskit-gym-mcp-server/pyproject.toml
[project]
name = "qiskit-gym-mcp-server"
version = "0.1.0"
description = "MCP server for qiskit-gym reinforcement learning quantum circuit synthesis"
readme = "README.md"
requires-python = ">=3.10,<3.14"

dependencies = [
    "fastmcp>=2.12.4",
    "qiskit-gym>=0.3.0",
    "qiskit>=1.3.0",
    "nest-asyncio>=1.5.0",
    "qiskit-mcp-server",
    "pydantic>=2.0.0",
    "qiskit-ibm-runtime>=0.34.0",  # For fake backends (exact IBM topologies without credentials)
]

[tool.uv.sources]
qiskit-mcp-server = { workspace = true }

[project.optional-dependencies]
dev = [
    "ruff>=0.11.13",
    "mypy>=1.15.0"
]

test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "pytest-cov>=4.1.0",
    "bandit[toml]>=1.9.2"
]

examples = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "python-dotenv>=1.0.0",
]

all = [
    "qiskit-gym-mcp-server[dev,test,examples]",
]

[[project.authors]]
name = "Qiskit Community"
email = "qiskit@us.ibm.com"

[project.license]
text = "Apache-2.0"

keywords = ["quantum", "qiskit", "ibm", "mcp", "ai", "quantum-computing", "reinforcement-learning", "circuit-synthesis"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/Qiskit/mcp-servers"
Repository = "https://github.com/Qiskit/mcp-servers"
Documentation = "https://github.com/Qiskit/mcp-servers/tree/main/qiskit-gym-mcp-server#readme"
"Bug Tracker" = "https://github.com/Qiskit/mcp-servers/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qiskit_gym_mcp_server"]
include = ["src/qiskit_gym_mcp_server/py.typed"]

[project.scripts]
qiskit-gym-mcp-server = "qiskit_gym_mcp_server:main"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "mypy>=1.15.0",
    "bandit[toml]>=1.9.2",
]
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "pytest-cov>=4.1.0",
]
examples = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "python-dotenv>=1.0.0",
]

# Bandit security configuration
[tool.bandit]
skips = ["B101"]  # Skip assert_used check in tests
exclude_dirs = ["tests", ".venv", "build", "dist"]
